模块更新，主题更换

+++ Src/GMS.Web.Admin/Areas/Account/Controllers/UserController.cs
@@ -13,37 +13,24 @@ namespace GMS.Web.Admin.Areas.Account.Controllers
     [Permission(EnumBusinessPermission.AccountManage_User)]
     public class UserController : AdminControllerBase
     {
-        //
-        // GET: /Account/User/
-
         public ActionResult Index(UserRequest request)
         {
             var result = this.AccountService.GetUserList(request);
             return View(result);
         }
-
-        //
-        // GET: /Account/User/Create
-
         public ActionResult Create()
         {
             var roles = this.AccountService.GetRoleList();
             this.ViewBag.RoleIds = new SelectList(roles, "ID", "Name");
-            
             var model = new User();
-            //model.Password = "111111";
-            return View("Edit", model);
+            return View(model);
         }
-
-        //
-        // POST: /Account/User/Create
-
         [HttpPost]
         public ActionResult Create(FormCollection collection)
         {
             var model = new User();
             this.TryUpdateModel<User>(model);
-            model.Password = Encrypt.MD5(model.Password);
+            model.Password = Encrypt.MD5(model.NewPassword);
             try
             {
                 this.AccountService.SaveUser(model);
@@ -51,48 +38,44 @@ namespace GMS.Web.Admin.Areas.Account.Controllers
             catch (BusinessException e)
             {
                 this.ModelState.AddModelError(e.Name, e.Message);
-
                 var roles = this.AccountService.GetRoleList();
                 this.ViewBag.RoleIds = new SelectList(roles, "ID", "Name");
-
-                return View("Edit", model);
+                return View(model);
             }
-
-            return this.RefreshParent();
+            return this.RedirectToAction("Index");
         }
-
-        //
-        // GET: /Account/User/Edit/5
-
         public ActionResult Edit(int id)
         {
             var model = this.AccountService.GetUser(id);
-
-             var roles = this.AccountService.GetRoleList();
-             this.ViewBag.RoleIds = new SelectList(roles, "ID", "Name", string.Join(",", model.Roles.Select(r => r.ID)));
-
+            var roles = this.AccountService.GetRoleList();
+            this.ViewBag.RoleIds = new SelectList(roles, "ID", "Name", string.Join(",", model.Roles.Select(r => r.ID)));
             return View(model);
         }
-
-        //
-        // POST: /Account/User/Edit/5
-
         [HttpPost]
-        public ActionResult Edit(int id, FormCollection collection)
+        public ActionResult Edit(int ID,FormCollection formCollection)
         {
-            var model = this.AccountService.GetUser(id);
+            var model = this.AccountService.GetUser(ID);
             this.TryUpdateModel<User>(model);
-            this.AccountService.SaveUser(model);
-
-            return this.RefreshParent();
+            if (!string.IsNullOrEmpty(model.NewPassword_NoRequired))
+                model.Password = Encrypt.MD5(model.NewPassword_NoRequired);
+            try
+            {
+                this.AccountService.SaveUser(model);
+            }
+            catch (BusinessException e)
+            {
+                this.ModelState.AddModelError(e.Name, e.Message);
+                var roles = this.AccountService.GetRoleList();
+                this.ViewBag.RoleIds = new SelectList(roles, "ID", "Name", string.Join(",", model.Roles.Select(r => r.ID)));
+                return View(model);
+            }
+            return this.RedirectToAction("Index");
         }
-
-        // POST: /Account/User/Delete/5
-
         [HttpPost]
         public ActionResult Delete(List<int> ids)
         {
-            this.AccountService.DeleteUser(ids);
+            if (ids != null)
+                this.AccountService.DeleteUser(ids);
             return RedirectToAction("Index");
         }
     }
